universe.py
```python
class Universe:
    def __init__(self, id, properties):
        """
        Initialize a new Universe instance.

        :param id: The unique identifier for this universe.
        :param properties: A dictionary containing the properties of this universe.
        """
        self.id = id
        self.properties = properties

    def simulate(self):
        """
        Simulate the universe. This method should be overridden by subclasses to provide
        specific simulation functionality.
        """
        raise NotImplementedError("Subclasses must implement this method.")

    def get_properties(self):
        """
        Get the properties of this universe.

        :return: A dictionary containing the properties of this universe.
        """
        return self.properties

    def set_properties(self, properties):
        """
        Set the properties of this universe.

        :param properties: A dictionary containing the new properties for this universe.
        """
        self.properties = properties
```
